Implement Author management using Artisan commands
I chose to have some fun and implement author CRUD using the Artisan console. Although custom console commands are something I rarely find the opportunity to use in day-to-day projects, it can't be understated how useful it is that Laravel provides this functionality out of the box.

You can `author:list` to see all registered authors, `author:create` to interactively create a new author, `author:reset-password {id}` to reset an author's password, and `author:delete {id}` to delete an author.