Implement Post model and API endpoints
I created a simple idea of what a post is, containing a title, body, author, and a set of tags.

I chose to take some advantage of MongoDB here and store the relevant information about the author (currently only name) within the post document itself rather than building out a relationship. I did the same thing with tags, making them free-form and storing them directly in the post rather than building a many-to-many relationship.

I was hoping to expand a bit on the idea of tags and include the ability to filter posts by tags, etc. but ran short on time. I'm hoping to work on implementing that this weekend.