Implement Vue.js single-page application
Aside from the login page, I went with a Vue.js SPA for reader and author functionality. Readers can browse all posts on the front page, and click into individual articles. An author, once logged in, can compose new posts as well as edit or delete existing posts.

A limited subset of Markdown is also available to authors to style their posts with. I used a few greate packages on the frontend including TailwindCSS, vue-markdown, vue-moment, and of course vue-router and vuex.